// src/app/model/list-rekening.ts
export interface listRekening {
  bankCode: any;
  bankName: any;
  bankBiCode: any;
  bankCity: any;
  bankBranch: any;
  accountNo: any;
  accountName: any;
  accountStatus: any;
  rekeningCode: any;
}

export interface listBank {
  bankCode: any;
  bankName: any;
}

export interface listCabangBank {
  bankBiCode: any;
  bankCity: any;
  bankBranch: any;
  bankCabang: any;
}